@@ -3,6 +3,7 @@ package Algorithms.Graph.HGA;
 
 import Algorithms.Graph.Hungarian;
 import Algorithms.Graph.NBM;
+import DS.Matrix.StatisticsMatrix;
 import DS.Network.Graph;
 import DS.Matrix.SimMat;
 import DS.Network.UndirectedGraph;
@@ -13,6 +14,7 @@ import com.aparapi.Range;
 import com.google.common.primitives.Doubles;
 import com.google.common.primitives.Floats;
 import org.apache.commons.io.FileUtils;
+import org.ejml.data.DMatrixRMaj;
 import org.jblas.DoubleMatrix;
 import org.jgrapht.alg.util.Pair;
 
@@ -47,7 +49,7 @@ public class HGA<V,E> {
     // parameters
     private boolean forcedMappingForSame;
     private double hAccount;
-    protected float bioFactor;
+    protected double bioFactor;
     private double edgeScore = 1.;
     private int h = 5;
     //---------------mapping result(best mapping)-------------
@@ -57,7 +59,7 @@ public class HGA<V,E> {
     private double PS_res;
     private double EC_res;
     private double score_res;
-    private DoubleMatrix matrix_res;
+    private StatisticsMatrix matrix_res;
     //---------------mapping for iteration---------
     public HashMap<V,V> mapping;
     private double PE;
@@ -67,7 +69,7 @@ public class HGA<V,E> {
     private double score;
     //---------------mapping for iteration---------
     private final SimMat<V> originalMat;
-    private Stack<DoubleMatrix> stackMat;
+    private Stack<StatisticsMatrix> stackMat;
     private Stack<Double> stackScore;
 
     //----------limit-----
@@ -88,7 +90,7 @@ public class HGA<V,E> {
 
     /**
      * Step 1:
-     * using homologous coefficients of proteins
+     * Initialize with the homologous coefficients of proteins
      * computed by alignment algorithms for PINs
      *
      * @param udG1                 graph1
@@ -97,11 +99,13 @@ public class HGA<V,E> {
      * @param nodalFactor          nodal compared with topological effect
      * @param forcedMappingForSame whether force mapping
      * @param hAccount             hungarian matrix account
+     * @param tolerance            the limit to check whether the matrix has converged
+     *
      */
     public HGA(SimMat<V> simMat,
                UndirectedGraph<V,E> udG1,
                UndirectedGraph<V,E> udG2,
-               float nodalFactor, boolean forcedMappingForSame, double hAccount, double tolerance) throws IOException {
+               double nodalFactor, boolean forcedMappingForSame, double hAccount, double tolerance) throws IOException {
 
         this.udG1 = udG1;
         this.udG2 = udG2;
@@ -314,7 +318,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
         // https://docs.oracle.com/javase/tutorial/collections/streams/parallelism.html
         // similarity matrix after the neighborhood adjustment
         SimMat<V> preSimMat = simMat.dup();
-        sumPreSimMat = preSimMat.getMat().sum();
+        sumPreSimMat = preSimMat.getMat().elementSum();
         // when index graph nodes scale is less than LIMIT then HGA uses parallel CPU instead
         // && nodes1.size() > LimitOfIndexGraph
         if (GPU) {
@@ -344,11 +348,11 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
         // so (i,j) in simMat -> int[] neighbors = nei_x[start_x[i],start_x[i+1]) and nei_y[start_y[j],start_y[j+1])
         // int[] non-neighbors = [0,n-1]-nei_x[start_x[i],start_x[i+1]) and [0,m-1]-nei_y[start_y[j],start_y[j+1])
         // result to be polished
-        float[] out =  Floats.toArray(Doubles.asList(simMat.getMat().data));
+        double[] out =  simMat.getMat().data();
         // preMat data
-        final float[] pre =  Floats.toArray(Doubles.asList(preMat.getMat().data));
+        final double[] pre =  preMat.getMat().data();
         // original data
-        final float[] ori =  Floats.toArray(Doubles.asList(originalMat.getMat().data));
+        final double[] ori =  originalMat.getMat().data();
         // initialize nei_x
         initNeighborToArray(nodes1, udG1, rowMap, nei_x, start_x);
         // initialize nei_y
@@ -362,7 +366,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
                 bioFactor);
         Range range = Range.create(nodes1.size()*nodes2.size(),1);
         kernel.execute(range).get(out);
-        simMat.getMat().data = Doubles.toArray(Floats.asList(out));
+        simMat.setData(out);
         kernel.dispose();
     }
 
@@ -502,17 +506,17 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
             if (iterCount > iterMax) {
                 return true;
             }
-            DoubleMatrix s1 = stackMat.get(1);
-            DoubleMatrix s = stackMat.peek();
+            StatisticsMatrix s1 = stackMat.get(1);
+            StatisticsMatrix s = stackMat.peek();
             // remove bottom which is the oldest for every iteration
-            DoubleMatrix s2 = stackMat.remove(0);
+            StatisticsMatrix s2 = stackMat.remove(0);
 
             double score = stackScore.peek();
             double score1 = stackScore.get(1);
             double score2 = stackScore.remove(0);
 
-            double dif_1 = s.sub(s1).normmax();
-            double dif_2 = s.sub(s2).normmax();
+            double dif_1 = s.minus(s1).elementMaxAbs();
+            double dif_2 = s.minus(s2).elementMaxAbs();
             logInfo("Iteration:" + iterCount + "\tdif_1 " + dif_1 + "\t" + "dif_2 " + dif_2 + "\nScore:" + "score1 " + score1
                     + "\t" + "score2 " + score2);
             return dif_1 < tolerance || dif_2 < tolerance ||
@@ -521,8 +525,8 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
         }
         // size = 2
         else if(stackMat.size() == 2){
-            DoubleMatrix s = stackMat.peek();
-            double dif = s.sub(stackMat.get(0)).normmax();
+            StatisticsMatrix s = stackMat.peek();
+            double dif = s.minus(stackMat.get(0)).elementMaxAbs();
             logInfo("Iteration:" + iterCount + "\tdif " + dif);
             return dif < tolerance;
         }
@@ -563,7 +567,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
             // step 2 score the mapping
             scoreMapping(this.mapping);
             // record
-            stackMat.push(this.simMat.getMat().dup());
+            stackMat.push(this.simMat.getMat().copy());
             stackScore.push(score);
             outDebug();
             // step 3 update based on mapped nodes
@@ -578,7 +582,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
             checkPassed = checkPassed(tolerance);
         } while (!checkPassed);
         // output result
-        logInfo("HGA mapping finish!With iteration "+iterCount+" times.");
+        logInfo("HGA mapping finish!With iteration "+this.iterCount+" times.");
         outPutResult();
 
     }
@@ -642,16 +646,14 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
         }
     }
 
-    public void outPutMatrix(DoubleMatrix mat, boolean isResult) {
+    public void outPutMatrix(StatisticsMatrix mat, boolean isResult) {
         logInfo("output matrix");
         String path = debugOutputPath + "matrix/";
         Vector<String> matrixVec = new Vector<>();
-        double[][] mat_ = mat.toArray2();
-        for (double[] doubles : mat_) {
-            for (int j = 0; j < mat_[0].length; j++) {
-                matrixVec.add(doubles[j] + " ");
+        for (int i = 0; i < mat.numRows(); i++) {
+            for (int j = 0; j < mat.numCols(); j++) {
+                matrixVec.add(mat.get(i,j) + " ");
             }
-            matrixVec.add("\n");
         }
         try {
             if (isResult) {
@@ -734,7 +736,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
 
         score_res = score;
         mappingResult = new HashMap<>(mapping);
-        matrix_res = simMat.getMat().dup();
+        matrix_res = simMat.getMat().copy();
     }
 
     private void initScores(double... scores) {
@@ -778,7 +780,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
         return score_res;
     }
 
-    public DoubleMatrix getMatrix_res() {
+    public StatisticsMatrix getMatrix_res() {
         return matrix_res;
     }
 
@@ -820,7 +822,7 @@ protected void addTopology(V node1, V node2, SimMat<V> preMat) {
     }
 
 
-    public void setBioFactor(float bioFactor) {
+    public void setBioFactor(double bioFactor) {
         assert (bioFactor >= 0 && bioFactor <= 1);
         this.bioFactor = bioFactor;
     }
