@@ -5,6 +5,7 @@ import DS.Matrix.SimMat;
 import DS.Network.UndirectedGraph;
 import IO.GraphFileReader;
 import IO.SimMatReader;
+import org.ejml.data.MatrixType;
 import org.jgrapht.graph.DefaultEdge;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -30,7 +31,7 @@ class HGASpc{
         udG2 = reader.readToUndirectedGraph( "src/test/java/resources/AlgTest/HGA/graph2.txt",false);
         SimMatReader<String> simMatReader = new SimMatReader<>(udG1.vertexSet(),udG2.vertexSet(),String.class);
         simMat = simMatReader.readToSimMat("src/test/java/resources/AlgTest/HGA/simMat.txt",true);
-        hga = new HGA<>(simMat, udG1, udG2, 0.5, true, 0.5, 0.01);
+        hga = new HGA<>(simMat, udG1, udG2, (float) 0.5, true, 0.5, 0.01);
     }
 
     @DisplayName("Greedily map")
